requirements.txt
sentence-transformers==3.1.0
djangorestframework==3.15.2
azure-storage-blob==12.23.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
elasticsearch==8.14.0
python-dotenv==1.0.1
scikit-learn==1.5.2
selenium==4.24.0
requests==2.32.3
notebook==7.2.2
torch==2.3.1+cpu
openai==1.35.7
numpy==1.26.4
tqdm==4.66.5
drf-yasg==1.21.7
pandas
streamlit
asyncpg
django-rest-swagger
markdown
django-filter
seaborn
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html